Convert HTML to PDF using Node.js module
For some reason, all of the HTML -> PDF renderers in JS and Ruby that I
can find don't work with the styles as they were. To make the rendering
work I've updated the styles to be in physical sizes as everything
becomes an A4 PDF anyway. It appears that the renderer is zoomed in as
using correct A4 measurements (21cm x 29.7cm) cuts a lot off.